DEBUG mode added & path.c, pl.c included (working)
I've included a DEBUG mode in hunter.c, so there's a #define at the
start of the file you can set to 0 or 1, if it's 1 then all the print
statement will occur, otherwise they will be silenced.

path.c and pl.c have been incorporated - let's us test if shortestPath
works (it does! well, it looks like it does). Now we can be a bit more
confident that hunter.c will be effective when running shortestMove.